requirements.txt
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
python_dateutil==2.8.2
scikit_learn==1.1.1
streamlit==1.10.0
yfinance==0.1.72
